Update and refactor Navbar
