+++ src/client/src/App.tsx
@@ -2,6 +2,8 @@ import { Route, Routes } from "react-router-dom";
 import Register from "./pages/Register";
 import RegisterSuccess from "./pages/Success";
 import {
+  ABOUT_US,
+  CONTACT_US,
   HOMEPAGE,
   LOGIN,
   REGISTER,
@@ -15,6 +17,8 @@ import Homepage from "./pages/Homepage";
 import RequireAuth from "./components/RequireAuth";
 import PersistLogin from "./components/PersistLogin";
 import Missing from "./pages/Missing";
+import AboutUs from "./pages/AboutUs";
+import ContactUs from "./pages/ContactUs";
 
 function App() {
   return (
@@ -26,7 +30,8 @@ function App() {
           <Route path={REGISTER_SUCCESS} element={<RegisterSuccess />} />
 
           <Route path={HOMEPAGE} element={<Homepage />} />
-
+          <Route path={ABOUT_US} element={<AboutUs />} />
+          <Route path={CONTACT_US} element={<ContactUs />} />
           <Route element={<RequireAuth allowedRole={"user"} />}>
             <Route path={VERIFY} element={<Verify />} />
           </Route>
