+++ src/client/src/components/Navbar/SignIn.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { LOGIN } from "../../constants/path";
+import { LOGIN, REGISTER } from "../../constants/path";
 
 const SignIn = () => {
   return (
@@ -11,12 +11,12 @@ const SignIn = () => {
       >
         Sign In
       </Link>
-      <a
-        href="#"
+      <Link
+        to={REGISTER}
         className="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium"
       >
         Join Us
-      </a>
+      </Link>
     </>
   );
 };
