+++ src/client/src/constants/path.ts
@@ -2,4 +2,8 @@ export const REGISTER = "/register";
 export const REGISTER_SUCCESS = "/register/success";
 export const LOGIN = "/login";
 export const VERIFY = "/verify-confirm";
+
+// Essential
 export const HOMEPAGE = "/";
+export const ABOUT_US = "/about-us";
+export const CONTACT_US = "/contact";
