+++ src/client/src/pages/AboutUs.tsx
@@ -1,0 +1,5 @@
+const AboutUs = () => {
+  return <div>AboutUs</div>;
+};
+
+export default AboutUs;
