+++ src/client/src/pages/ContactUs.tsx
@@ -1,0 +1,5 @@
+const ContactUs = () => {
+  return <div>ContactUs</div>;
+};
+
+export default ContactUs;
